third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-smart-braces-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks text editor smart braces functionality.


Running: testTypeBraceSequence
({[|]})

Running: testBraceOverride
({[]})|

Running: testQuotesToCloseStringLiterals
'Hello"'|

Running: testQuotesToCloseStringLiteralInsideLine
console.log("information"|);
